% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManipulation.R
\name{aggregateData}
\alias{aggregateData}
\title{aggregateData}
\usage{
aggregateData(trips, year, save_data = TRUE)
}
\arguments{
\item{trips}{the raw data}

\item{year}{the year of the raw predictions}

\item{save_data}{if TRUE the data are saved in a .RData format with the name "AggregatedData<year>"}
}
\value{

}
\description{
aggregate the raw Divvy_Bike_Data in a Data_Frame with aggregated parameter
}
\author{
Rémy Garnier & Camille Palmier
}
