% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDataSet.R
\name{BuildDataSet}
\alias{BuildDataSet}
\title{Build Data Set}
\usage{
BuildDataSet(year, raw_data, districts, Stations, save_data = FALSE)
}
\arguments{
\item{year}{the year}

\item{raw_data}{data coming from Aggregate Data}

\item{districts}{a Large SpatialPolygonDataFrame containing the different aera . The name of the column containing the area should be SLDLST}

\item{save_data}{if TRUE the data are saved in a .RData format with the name "SpecialDays<year>"}

\item{Stations.map}{a data frame containing the position of the stations (latitude, longitude)}
}
\value{
A data frame containing the following information

Time : the timestamp of the observation
Day : The Day of observation
Hour : The Hour of the observation
dow : the position of the day in the week (8 for holydays)
weekday : name of the day (or of the hollydays)
district : the subregion considered
nbE : number of bikes arriving in a station of the considered district during the period of 1 hour  
nbS : number of bikes leaving in a station of the considered district during the period of 1 hour 
diff : nbE - nbS
temp : temperature
pluvio : pluviometry
}
\description{
Build the full data set
}
\author{
Rémy Garnier & Camille Palmier
}
