% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeteo.R
\name{getMeteo}
\alias{getMeteo}
\title{getMeteo}
\usage{
getMeteo(year)
}
\arguments{
\item{year}{the year considerate}
}
\value{

}
\description{
get the meteological data of Chicago on a year
}
\author{
Rémy Garnier & Camille Palmier
}
