% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecialDays.R
\name{getSpecialDays}
\alias{getSpecialDays}
\title{getSpecialDays}
\usage{
getSpecialDays(year, save_data = TRUE)
}
\arguments{
\item{year}{the year considerate}

\item{save_data}{if TRUE the data are saved in a .RData format with the name "SpecialDays<year>"}
}
\value{

}
\description{
get a data frame with three colums :
first one for the day of the year ;
second one with the name of the week or of the special days ;
last one with the day of the week where 1 is monday, ... , 7 is sunday and 8 is special day.
}
\author{
Rémy Garnier & Camille Palmier
}
