% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeospatialData.R
\name{groupByGeospatialData}
\alias{groupByGeospatialData}
\title{groupByGeospatialData}
\usage{
groupByGeospatialData(data, area, stations, save_data = TRUE)
}
\arguments{
\item{data}{trips dataframe grouped by hour (see aggregateData for the format)}

\item{area}{a Large SpatialPolygonDataFrame containing the different aera . The name of the column containing the area should be SLDLST}

\item{stations}{a data frame containing the position of the stations (latitude, longitude)}

\item{save_data}{decide whether to save the data or not}
}
\value{

}
\description{
Group the station by geographical area
}
\author{
Rémy Garnier & Camille Palmier
}
