% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManipulation.R
\name{aggregateData}
\alias{aggregateData}
\title{aggregateData}
\usage{
aggregateData(trips, year)
}
\arguments{
\item{trips}{the raw data}

\item{year}{the predictions}
}
\value{

}
\description{
aggregate the raw Divvy_Bike_Data in a Data_Frame with aggregated parameter
}
\author{
Rémy Garnier & Camille Palmier
}
