% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeteo.R
\name{getMeteo}
\alias{getMeteo}
\title{getMeteo}
\usage{
getMeteo(year, save_data = TRUE)
}
\arguments{
\item{year}{the year considerate}

\item{save_data}{if TRUE the data are saved in a .RData format with the name "MeteoData<year>"}
}
\value{

}
\description{
get the meteological data of Chicago on a year
warning : internet connection is required
}
\author{
Rémy Garnier & Camille Palmier
}
